bugfix: don't get stuck in modalGamma loops when new node is already on branch
In the S4 tree for (A ∧ ¬□A)→□¬□A, (¬A ∧ ◇A)→□◇A |= ◇□A→□◇A the tree construction got stuck in a loop at step 57. Here (or in 60 etc.), it should try expanding the modalGamma formula with w2 instead of w1, but it kept trying w1 because the result of that modalGamma application is already on the branch, so it couldn't find it when looking if an instance term has already been used.