improve processing of alternative trees in backtracking
When reaching the depthLimit, I used to discard the current tree and switch to
one of its alternatives; when there were no alternatives left, I increased the
depthLimit and kept working with the last tree.

This is not ideal because the last tree in the list of alternatives is often not
the simplest, and we effectively never reconsider the earlier
alternatives. Restarting the tree construction from scratch whenever the
depthlimit is reached fixes this, but it creates a lot of redundancy.

This patch instead keeps the alternatives in memory even when depthlimit is
reached. When the limit is increased, we simply go back to the first alternative
and expand it some more.