don't re-apply seriality rule before other rules
The seriality rule has lowest priority to make sure that when a new
node wRv is added, we don't continue with vRu, uRs, sRt, ... if there's
anything else we could do instead.

The code in Prover.addAccessibilityRuleApplications for adding new
seriality applications assumed that branch.todoList is in order of
priority, and added the new application before the first entry with the
same priority. However, gamma and modalGamma add their re-application to
the end of todoList, no matter the priorities. So there can be todoLists
that begin with a seriality application followed by gamma/modalGamma
applications. The old code then added the new seriality application to
the front of the list, causing an infinite application of seriality
rules. Example: ◇(p∧□q)→◇(p∧◇q).

I now make sure that when a new rule application is added to
branch.todoList, it comes after any rules with greater priority.