make branch.id never part of other branch.id; closes #4
Tree.markUsedNodes only adds the current branch id to n.used if the id
is not already a substring of n.used. If the current branch has id 'b1'
and n.used contains 'b1234', the current branch id isn't added. This
caused branches to be pruned that shouldn't be pruned, as in issue #4 on
github. (The effect only showed up on the first run because on the
second run the initial branch id is no longer 1.)