fix regression: don't use tseitin transforms only for some formulas
When creating the clauses for input 'A <-> (A <-> B) |= A <-> B', the
tseitin transform for the premise is not used, so the transform for the
conclusion shouldn't use the tseitin variables ($1...) that were created
for the premise.

I now either use the tseitin transform for all formulas or for none.
Probably not maximally efficient, but it doesn't slow down any of my
tests.