fix order of [] and (Ax) processing
When expanding □∀x□∀y(□Fx∨□Gy)→(□∀x□Fx∨□∀x□Gx) in S4, we have a lot of
options of wR* for the outermost box in the antecedent; only one of them
leads to a closed tree. Previously, the prover was very slow at
iterating through the options because the modalGamma function put itself
at the very end of the todo queue after each application. Now we
immediately expand [] formulas with all available options.