fix pruneAlternatives
When checking for redundant alternatives among prover.alternatives, I
used to compare different trees by checking if each open branch on one
extends some open branch on another. This is problematic if the trees
contain a disjunction p v p (with the same disjunct on both sides). Any
branch that develops the left disjunct on tree1 then extends the
undeveloped right-hand branch on tree2, even if tree1 and tree2
otherwise explore very different strategies.

The present change helps with issue #23 on github, but it causes a
regression for Pelletier's problem 51, which was previously provable in
2508 steps and now appears to become unprovable in any reasonable time.
(Problems 49 and 52 also perform much worse.)