Restrict equality reasoner to individuals; fixes #31
I used to allow the equality reasoner to substitute individual variable
for world variables. But the skolem terms for the two types of variables
only keep track of instances of the same type. This caused bugs like
lrbs to only return terms for individuals, so that equations involving
world terms are never considered.